Refactored data loading to use torch's IterableDataset and DataLoader for multiprocessing-based data loading
