+++ components/train/run_local.sh
@@ -27,6 +27,7 @@ python3 src/train.py \
   --test_list=vox1_no_cuda.txt --train_list=vox2_no_cuda.txt \
   --test_path=vox1_no_cuda_feats.tar.gz --train_path=vox2_no_cuda_feats.tar.gz \
   --batch_size=5 --nSpeakers=2 --max_epoch=2 --test_interval=1 \
+  --n-data-loader-thread=5\
   $@
 
 ## full data catered to milo's local hardware
@@ -34,5 +35,6 @@ python3 src/train.py \
 #  --data-bucket=voxsrc-2020-voxceleb-v4 \
 #  --test_list=vox1_full.txt --train_list=vox2_full.txt \
 #  --test_path=vox1_full_feats.tar.gz --train_path=vox2_full_feats.tar.gz \
-#  --batch_size=400 --nSpeakers=2 --max_epoch=500 --test_interval=3\
+#  --batch_size=400 --nSpeakers=2 --max_epoch=500 --test_interval=3 \
+#  --n-data-loader-thread=5 \
 #  $@
